+++ AService/Controllers/BookStoreController.cs
@@ -1,14 +1,15 @@
-﻿using AService.Items;
+﻿using System.Web.Http;
+using System.Web.Http.Description;
+using AService.Items;
 using AService.Models;
-using Microsoft.AspNetCore.Mvc;
+
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Options;
 
 namespace AService.Controllers
 {
-	[Route("api/bs")]
-	[ApiController]
-	public class BookStoreController : Controller
+	[RoutePrefix("api/bs")]	
+	public class BookStoreController : ApiController
 	{
 		private readonly BookStoreContext _context;
 		private readonly IOptions<ServiceOptions> options;
@@ -19,20 +20,17 @@ namespace AService.Controllers
 			this.options = options;
 		}
 
-		// GET: api/bs
-		[HttpGet]
-		public async Task<ActionResult<IEnumerable<Book>>> GetItems()
-		{
-			if (_context.Books == null)
-			{
-				return NotFound();
-			}
+		// GET: api/bs		
+		[Route("")]
+		public async Task<IEnumerable<Book>> GetItems()
+		{			
 			return await _context.Books.ToListAsync();
 		}
 
 		// GET: api/bs/5
-		[HttpGet("{id}")]
-		public async Task<ActionResult<Book>> GetItem(long id)
+		[Route("{id:int}")]
+		[ResponseType(typeof(Book))]
+		public async Task<IHttpActionResult> GetItem(long id)
 		{
 			if (_context.Books == null)
 			{
@@ -45,78 +43,80 @@ namespace AService.Controllers
 				return NotFound();
 			}
 
-			return item;
+			return Ok(item);
 		}
 
-		// PUT: api/bs/5
-		// To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
-		[HttpPut("{id}")]
-		public async Task<IActionResult> PutItem(long id, Book item)
-		{
-			if (id != item.Id)
-			{
-				return BadRequest();
-			}
+		// https://learn.microsoft.com/en-us/archive/blogs/youssefm/writing-tests-for-an-asp-net-web-api-service
 
-			_context.Entry(item).State = EntityState.Modified;
+		//// PUT: api/bs/5
+		//// To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
+		//[HttpPut("{id}")]
+		//public async Task<IActionResult> PutItem(long id, Book item)
+		//{
+		//	if (id != item.Id)
+		//	{
+		//		return BadRequest();
+		//	}
 
-			try
-			{
-				await _context.SaveChangesAsync();
-			}
-			catch (DbUpdateConcurrencyException)
-			{
-				if (!ItemExists(id))
-				{
-					return NotFound();
-				}
-				else
-				{
-					throw;
-				}
-			}
+		//	_context.Entry(item).State = EntityState.Modified;
 
-			return NoContent();
-		}
+		//	try
+		//	{
+		//		await _context.SaveChangesAsync();
+		//	}
+		//	catch (DbUpdateConcurrencyException)
+		//	{
+		//		if (!ItemExists(id))
+		//		{
+		//			return NotFound();
+		//		}
+		//		else
+		//		{
+		//			throw;
+		//		}
+		//	}
 
-		// POST: api/bs
-		// To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
-		[HttpPost]
-		public async Task<ActionResult<Item>> PostItem(Book item)
-		{
-			if (_context.Books == null)
-			{
-				return Problem("Entity set 'ItemContext.Items'  is null.");
-			}
-			_context.Books.Add(item);
-			await _context.SaveChangesAsync();
+		//	return NoContent();
+		//}
 
-			return CreatedAtAction(nameof(GetItem), new { id = item.Id }, item);
-		}
+		//// POST: api/bs
+		//// To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
+		//[HttpPost]
+		//public async Task<ActionResult<Item>> PostItem(Book item)
+		//{
+		//	if (_context.Books == null)
+		//	{
+		//		return Problem("Entity set 'ItemContext.Items'  is null.");
+		//	}
+		//	_context.Books.Add(item);
+		//	await _context.SaveChangesAsync();
 
-		// DELETE: api/bs/5
-		[HttpDelete("{id}")]
-		public async Task<IActionResult> DeleteItem(long id)
-		{
-			if (_context.Books == null)
-			{
-				return NotFound();
-			}
-			var item = await _context.Books.FindAsync(id);
-			if (item == null)
-			{
-				return NotFound();
-			}
+		//	return CreatedAtAction(nameof(GetItem), new { id = item.Id }, item);
+		//}
 
-			_context.Books.Remove(item);
-			await _context.SaveChangesAsync();
+		//// DELETE: api/bs/5
+		//[HttpDelete("{id}")]
+		//public async Task<IActionResult> DeleteItem(long id)
+		//{
+		//	if (_context.Books == null)
+		//	{
+		//		return NotFound();
+		//	}
+		//	var item = await _context.Books.FindAsync(id);
+		//	if (item == null)
+		//	{
+		//		return NotFound();
+		//	}
 
-			return NoContent();
-		}
+		//	_context.Books.Remove(item);
+		//	await _context.SaveChangesAsync();
 
-		private bool ItemExists(long id)
-		{
-			return (_context.Books?.Any(e => e.Id == id)).GetValueOrDefault();
-		}
+		//	return NoContent();
+		//}
+
+		//private bool ItemExists(long id)
+		//{
+		//	return (_context.Books?.Any(e => e.Id == id)).GetValueOrDefault();
+		//}
 	}
 }
